Made constraint data structures pluggable.
Factored out interface for constraints. Current implementation: NaiveConstraint.
Preparing for a more efficient one.
