+++ src/dotty/tools/dotc/core/ConstraintRunInfo.scala
@@ -1,0 +1,16 @@
+package dotty.tools.dotc
+package core
+
+import Contexts._, config.Printers._
+
+trait ConstraintRunInfo { self: RunInfo =>
+  private var maxSize = 0
+  private var maxConstraint: Constraint = _
+  def recordConstraintSize(c: Constraint, size: Int) =
+    if (size > maxSize) {
+      maxSize = size
+      maxConstraint = c
+    }
+  def printMaxConstraint()(implicit ctx: Context) =
+    if (maxSize > 0) typr.println(s"max constraint = ${maxConstraint.show}")
+}
